+++ engine/src/main/java/dev/gamekit/ui/widgets/Center.java
@@ -2,9 +2,13 @@ package dev.gamekit.ui.widgets;
 
 import dev.gamekit.ui.enums.Alignment;
 
-/** A {@link Parent} which centers its single child */
+/** An {@link Align} which centers its single child */
 public class Center extends Align {
-  public Center(Widget child) {
-    super(child, Alignment.CENTER);
+  protected Center(Widget child) {
+    super(Alignment.CENTER, child);
+  }
+
+  public static Center create(Widget child) {
+    return new Center(child);
   }
 }